// src/interfaces/Vehicles.ts
export default interface Vehicles {
  name: string;
  model: string;
  manufacturer: string;
  cost_in_credits: string;
  length: string;
  max_atmosphering_speed: string;
  crew: string;
  passengers: string;
  cargo_capacity: string;
  url: string;
  consumables: string;
  vehicle_class: string;
  pilots: Array<string>;
  films: Array<string>;
}